Add ordering
Add command line options

Add LUT format file extensions

Add missing override

Remove source logging

Add convenience methods

Formatting

Documentation fixes

Removed play-fltk example

Add QImage overlay

Flip image

Flip image

Add methods for getting output device information

Change methods to signals

Add qualifier

If the overlay is null reset the overlay texture

Render the overlay at the viewport size

Size fix

Add support for QImage::Format_ARGB4444_Premultiplied

Blending fix

Blending test change

Revert change

QImage copy fix

Revert change

Take ownership of the QImage

Refactoring

Updated CMakeLists.txt files.

Fixed Imath_LIBRARIES variable referring to non-existant Imath_Imath_LIBRARY variable.

Added a getShader() member function in tl::gl::Render.

This is to get access to all the shaders in the std::map, like the "text" shader.
