+++ etc/SuperBuild/CMakeLists.txt
@@ -1,7 +1,9 @@
 cmake_minimum_required(VERSION 3.19)
 cmake_policy(SET CMP0091 NEW)
 cmake_policy(SET CMP0114 NEW)
-cmake_policy(SET CMP0135 NEW)
+if( POLICY CMP0135 )
+  cmake_policy(SET CMP0135 NEW)
+endif()
 
 project(tlRender-SuperBuild)
