+++ etc/SuperBuild/cmake/Modules/BuildtlRender.cmake
@@ -3,7 +3,7 @@ include(ExternalProject)
 set(tlRender_ARGS
     ${TLRENDER_EXTERNAL_ARGS}
     -DTLRENDER_ENABLE_MMAP=${TLRENDER_ENABLE_MMAP}
-	-DTLRENDER_ENABLE_COVERAGE=${TLRENDER_ENABLE_COVERAGE}
+    -DTLRENDER_ENABLE_COVERAGE=${TLRENDER_ENABLE_COVERAGE}
     -DTLRENDER_ENABLE_PYTHON=${TLRENDER_ENABLE_PYTHON}
     -DTLRENDER_BUILD_GL=${TLRENDER_BUILD_GL}
     -DTLRENDER_BUILD_BMD=${TLRENDER_BUILD_BMD}
