+++ lib/tlBakeApp/App.cpp
@@ -75,14 +75,10 @@ namespace tl
                         "The output file.")
                 },
         {
-            app::CmdLineValueOption<int64_t>::create(
-                _options.startFrame,
-                { "-startFrame", "-sf" },
-                "Start frame."),
-            app::CmdLineValueOption<int64_t>::create(
-                _options.endFrame,
-                { "-endFrame", "-ef" },
-                "End frame."),
+            app::CmdLineValueOption<otime::TimeRange>::create(
+                _options.inOutRange,
+                { "-inOutRange" },
+                "Set the in/out points range."),
             app::CmdLineValueOption<imaging::Size>::create(
                 _options.renderSize,
                 { "-renderSize", "-rs" },
@@ -108,7 +104,17 @@ namespace tl
             app::CmdLineValueOption<std::string>::create(
                 _options.colorConfigOptions.view,
                 { "-colorView", "-cv" },
-                "View color space.")
+                "View color space."),
+            app::CmdLineValueOption<std::string>::create(
+                _options.lutOptions.fileName,
+                { "-lut" },
+                "LUT file name."),
+            app::CmdLineValueOption<timeline::LUTOrder>::create(
+                _options.lutOptions.order,
+                { "-lutOrder" },
+                "LUT operation order.",
+                string::Format("{0}").arg(_options.lutOptions.order),
+                string::join(timeline::getLUTOrderLabels(), ", "))
         });
         }
 
@@ -152,12 +158,11 @@ namespace tl
             _print(string::Format("Timeline speed: {0}").arg(_duration.rate()));
 
             // Time range.
-            auto startTime = _options.startFrame >= 0 ?
-                otime::RationalTime(_options.startFrame, _duration.rate()) :
-                otime::RationalTime(0, _duration.rate());
-            _range = _options.endFrame >= 0 ?
-                otime::TimeRange::range_from_start_end_time_inclusive(startTime, otime::RationalTime(_options.endFrame, _duration.rate())) :
-                otime::TimeRange::range_from_start_end_time(startTime, startTime + _duration);
+            _range = _options.inOutRange != time::invalidTimeRange ?
+                _options.inOutRange :
+                otime::TimeRange::range_from_start_end_time(
+                    otime::RationalTime(0, _duration.rate()), 
+                    _duration);
             _currentTime = _range.start_time();
             _print(string::Format("Frame range: {0}-{1}").arg(_range.start_time().value()).arg(_range.end_time_inclusive().value()));
 
@@ -287,6 +292,7 @@ namespace tl
 
             // Render the video.
             _render->setColorConfig(_options.colorConfigOptions);
+            _render->setLUT(_options.lutOptions);
             _render->begin(_renderSize);
             const auto videoData = _timeline->getVideo(_timeline->getGlobalStartTime() + _currentTime).get();
             _render->drawVideo({ videoData }, { math::BBox2i(0, 0, _renderSize.w, _renderSize.h ) });
