+++ lib/tlIO/OpenEXRRead.cpp
@@ -130,7 +130,7 @@ namespace tl
                 {
                     // Open the file.
 #if defined(TLRENDER_ENABLE_MMAP)
-                // \bug https://lists.aswf.io/g/openexr-dev/message/43
+                    // \bug https://lists.aswf.io/g/openexr-dev/message/43
                     _s.reset(new MemoryMappedIStream(fileName.c_str()));
                     _f.reset(new Imf::InputFile(*_s.get()));
 #else // TLRENDER_ENABLE_MMAP
