+++ tests/tlTimelineTest/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(HEADERS
 	ColorConfigOptionsTest.h
 	IRenderTest.h
+	LUTOptionsTest.h
     TimelinePlayerTest.h
     TimelineTest.h
     TimelineUtilTest.h)
@@ -8,6 +9,7 @@ set(HEADERS
 set(SOURCE
 	ColorConfigOptionsTest.cpp
 	IRenderTest.cpp
+	LUTOptionsTest.cpp
     TimelinePlayerTest.cpp
     TimelineTest.cpp
     TimelineUtilTest.cpp)
