+++ tests/tlTimelineTest/LUTOptionsTest.h
@@ -1,0 +1,24 @@
+// SPDX-License-Identifier: BSD-3-Clause
+// Copyright (c) 2021-2022 Darby Johnston
+// All rights reserved.
+
+#pragma once
+
+#include <tlTestLib/ITest.h>
+
+namespace tl
+{
+    namespace timeline_tests
+    {
+        class LUTOptionsTest : public tests::ITest
+        {
+        protected:
+            LUTOptionsTest(const std::shared_ptr<system::Context>&);
+
+        public:
+            static std::shared_ptr<LUTOptionsTest> create(const std::shared_ptr<system::Context>&);
+
+            void run() override;
+        };
+    }
+}
